cxx-qt-lib: split qmap.h and qhash.h into one for each type
