+++ crates/cxx-qt-lib/include/core/qhash/qhash.h
@@ -1,0 +1,11 @@
+// clang-format off
+// SPDX-FileCopyrightText: 2025 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+// clang-format on
+// SPDX-FileContributor: Joshua Booth <joshua.n.booth@gmail.com>
+//
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+#pragma once
+
+#include "qhash_QString_QVariant.h"
+#include "qhash_i32_QByteArray.h"
