+++ crates/cxx-qt-lib/include/core/qhash/qhash_i32_QByteArray.h
@@ -1,0 +1,13 @@
+// clang-format off
+// SPDX-FileCopyrightText: 2025 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+// clang-format on
+// SPDX-FileContributor: Joshua Booth <joshua.n.booth@kdab.com>
+//
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+#pragma once
+#include "qhash_private.h"
+
+#include <QtCore/QByteArray>
+
+using QHash_i32_QByteArray = QHash<::std::int32_t, QByteArray>;
