+++ crates/cxx-qt-lib/include/core/qhash/qhash_private.h
@@ -6,10 +6,9 @@
 // SPDX-License-Identifier: MIT OR Apache-2.0
 #pragma once
 
-#include <QtCore/QMap>
+#include <cstdint>
 
-#include <QtCore/QString>
-#include <QtCore/QVariant>
+#include <QtCore/QHash>
 
 #include "rust/cxx.h"
 
@@ -19,74 +18,77 @@ namespace rust {
 
 // This has static asserts in the cpp file to ensure this is valid.
 template<typename K, typename V>
-struct IsRelocatable<QMap<K, V>> : ::std::true_type
+struct IsRelocatable<QHash<K, V>> : ::std::true_type
 {};
 
 } // namespace rust
 
 namespace rust {
 namespace cxxqtlib1 {
-namespace qmap {
+namespace qhash {
 
 template<typename K, typename V>
 ::rust::isize
-qmapLen(const QMap<K, V>& m) noexcept;
+qhashLen(const QHash<K, V>& h) noexcept;
 
 template<typename K, typename V>
 V
-qmapGetOrDefault(const QMap<K, V>& m, const K& key) noexcept
+qhashGetOrDefault(const QHash<K, V>& h, const K& key) noexcept
 {
   // Qt 6 returns a T and Qt 5 returns an const T
   // so we need to define our own method here for CXX
-  return m.value(key);
+  return h.value(key);
 }
 
 template<typename K, typename V>
 const K&
-qmapGetUncheckedKey(const QMap<K, V>& m, ::rust::isize pos) noexcept
+qhashGetUncheckedKey(const QHash<K, V>& h, ::rust::isize pos) noexcept
 {
-  Q_ASSERT(pos < qmapLen(m));
+  Q_ASSERT(pos < qhashLen(h));
   Q_ASSERT(pos >= 0);
-  auto it = m.cbegin();
+  auto it = h.cbegin();
   ::std::advance(it, pos);
   return it.key();
 }
 
 template<typename K, typename V>
 const V&
-qmapGetUncheckedValue(const QMap<K, V>& m, ::rust::isize pos) noexcept
+qhashGetUncheckedValue(const QHash<K, V>& h, ::rust::isize pos) noexcept
 {
-  Q_ASSERT(pos < qmapLen(m));
+  Q_ASSERT(pos < qhashLen(h));
   Q_ASSERT(pos >= 0);
-  auto it = m.cbegin();
+  auto it = h.cbegin();
   ::std::advance(it, pos);
   return it.value();
 }
 
 template<typename K, typename V>
 void
-qmapInsert(QMap<K, V>& m, const K& key, const V& value) noexcept
+qhashInsert(QHash<K, V>& h, const K& key, const V& value) noexcept
 {
-  m.insert(key, value);
+  h.insert(key, value);
 }
 
 template<typename K, typename V>
 ::rust::isize
-qmapLen(const QMap<K, V>& m) noexcept
+qhashLen(const QHash<K, V>& h) noexcept
 {
-  // Qt has an int as the QMap::size_type
-  return static_cast<::rust::isize>(m.size());
+  // Qt 6 returns a qsizetype and Qt 5 returns an int
+  return static_cast<::rust::isize>(h.size());
 }
 
 template<typename K, typename V>
 bool
-qmapRemove(QMap<K, V>& m, const K& key) noexcept
+qhashRemove(QHash<K, V>& h, const K& key) noexcept
 {
-  return m.remove(key) >= 1;
+  // Qt 6 returns a bool and Qt 5 returns an int
+#if (QT_VERSION >= QT_VERSION_CHECK(6, 0, 0))
+  return h.remove(key);
+#else
+  return h.remove(key) >= 1;
+#endif
 }
 
 }
 }
 }
-
-using QMap_QString_QVariant = QMap<QString, QVariant>;
