+++ crates/cxx-qt-lib/include/core/qmap/qmap_QString_QVariant.h
@@ -1,0 +1,15 @@
+// clang-format off
+// SPDX-FileCopyrightText: 2025 Klarälvdalens Datakonsult AB, a KDAB Group company <info@kdab.com>
+// clang-format on
+// SPDX-FileContributor: Joshua Booth <joshua.n.booth@gmail.com>
+//
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+#pragma once
+
+#include "qmap_private.h"
+
+#include <QtCore/QString>
+#include <QtCore/QVariant>
+
+using QMap_QString_QVariant = QMap<QString, QVariant>;
