+++ crates/cxx-qt-lib/src/core/qmap/qmap_qstring_qvariant.rs
@@ -13,7 +13,7 @@ pub mod ffi {
         include!("cxx-qt-lib/qvariant.h");
         type QVariant = crate::QVariant;
 
-        include!("cxx-qt-lib/qmap.h");
+        include!("cxx-qt-lib/qmap_QString_QVariant.h");
         type QMap_QString_QVariant = crate::QMap<super::QMapPair_QString_QVariant>;
     }
